+++ services/mongo_service.go
@@ -8,30 +8,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
 )
-
-// NewMongoConnection - creates a MongoConnection instance
-func NewMongoConnection(c string) (*mongo.Client, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	client, err := mongo.Connect(ctx, options.Client().ApplyURI(c))
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = client.Ping(ctx, readpref.Primary())
-
-	if err != nil {
-		log.Fatal("Cannot connect to mongo database")
-		return nil, err
-	}
-
-	log.Info("Connected to database")
-	return client, nil
-}
 
 // MongoService - a service to interact with a task
 type MongoService struct {
